requirements.txt
mangum
fastapi
uvicorn

pydantic
openai

python-dotenv
